requirements.txt
# Core
python-dotenv==1.0.1
requests==2.32.3
loguru==0.7.2
aiohttp==3.9.1
beautifulsoup4==4.12.2

# Speech to Text
sounddevice==0.4.7
faster-whisper==1.0.3
numpy>=1.26.0

# Text to Speech
pyttsx3==2.98

# AI APIs
google-generativeai==0.8.3
openai==1.6.1

# PC Control
pyautogui==0.9.54
psutil==5.9.6

# Free APIs Integration
duckduckgo-search==6.3.7
wikipedia==1.4.0
newsapi-python==0.2.7
pyowm==3.3.0
spotipy==2.22.1
youtube-search-python==1.6.6
twilio==8.11.0
telebot==0.0.5
pyTelegramBotAPI==4.14.0
instagram-private-api==1.6.0
facebook-sdk==3.1.0

# Data & Analytics
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0

# Optional (for later phases)
customtkinter==5.2.2
opencv-python==4.10.0.84
pycaw==20251023
schedule==1.2.0
qrcode==7.4.2
Pillow==10.1.0